app/backend/tests/requirements-test.txt
# 테스트 전용 의존성
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.25.2
respx==0.20.2

# 기존 의존성 (requirements.txt에서 가져옴)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
undetected-chromedriver==3.5.4
selenium==4.15.2
beautifulsoup4==4.12.2
reportlab==4.0.7
Pillow==10.1.0
pydantic==2.5.0
python-barcode==0.15.1
qrcode==7.4.2
slowapi==0.1.9
redis==5.0.1
python-dotenv==1.0.0